created challenge controller
